Improved rating system

+++ django/farmapp/models.py
@@ -228,6 +228,6 @@ class Review(models.Model):
     user_id = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE,related_name='user_id')
     customer = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE,related_name='customer')
     cart_id = models.ForeignKey(Cart, on_delete=models.CASCADE)
-    rating = models.PositiveIntegerField(default=0)
+    rating = models.FloatField(default=0)
     review = models.TextField()
     timestamp = models.DateTimeField(default=django.utils.timezone.now)
